Move withLockedObservation to separate class
In preparation for moving some biomass-related update functions to `BiomassStore`,
split the observation locking function out into its own class that can be a
dependency of both `BiomassStore` and `ObservationStore`. Otherwise we would have
to make `BiomassStore` depend on `ObservationStore` which we try to avoid.
