+++ src/test/kotlin/com/terraformation/backend/tracking/db/biomassStore/BaseBiomassStoreTest.kt
@@ -9,6 +9,7 @@ import com.terraformation.backend.db.default_schema.OrganizationId
 import com.terraformation.backend.db.tracking.PlantingSiteId
 import com.terraformation.backend.mockUser
 import com.terraformation.backend.tracking.db.BiomassStore
+import com.terraformation.backend.tracking.db.ObservationLocker
 import io.mockk.every
 import org.junit.jupiter.api.BeforeEach
 
@@ -17,7 +18,7 @@ abstract class BaseBiomassStoreTest : DatabaseTest(), RunsAsUser {
 
   protected val eventPublisher = TestEventPublisher()
   protected val store: BiomassStore by lazy {
-    BiomassStore(dslContext, eventPublisher, ParentStore(dslContext))
+    BiomassStore(dslContext, eventPublisher, ObservationLocker(dslContext), ParentStore(dslContext))
   }
 
   protected lateinit var organizationId: OrganizationId
